Refactor to keep all dialog code in one file + scroll correct containers.
